+++ crates/zng/build.rs
@@ -12,4 +12,11 @@ fn main() {
         crash_handler: { all(feature = "crash_handler", not(any(android, wasm))) },
         trace_recorder: { all(feature = "trace_recorder", not(any(android, wasm))) },
     }
+
+    // TODO(breaking) remove feature
+    if cfg!(feature = "view_bundle_licenses") && !cfg!(feature = "_all_features") {
+        println!(
+            "cargo::warning=feature \"view_bundle_licenses\" is deprecated, zng-view licenses are already collected when compiling with \"view\" and using `collect_cargo_about`"
+        );
+    }
 }